+++ src/commands/lastfm/mod.rs
@@ -1,0 +1,9 @@
+use serenity::framework::standard::macros::group;
+
+pub mod np;
+
+use self::np::*;
+
+#[group]
+#[commands(np)]
+pub struct LastFm;